all 3LED functions are working, the sleep wake up feature not finished yet.
