+++ include/sleep.h
@@ -1,0 +1,8 @@
+#ifndef GLOBALS_H
+#define GLOBALS_H
+#pragma once
+
+void checkIdleAndSleep();
+
+
+#endif
